bootloader: Fix some type warnings.
Change `int` to `size_t` in some places.

+++ bootloader/src/pyi_archive.c
@@ -562,7 +562,7 @@ char *
 pyi_arch_get_option(const ARCHIVE_STATUS * status, char * optname)
 {
     /* TODO: option-cache? */
-    int optlen;
+    size_t optlen;
     TOC *ptoc = status->tocbuff;
 
     optlen = strlen(optname);
